Move VerifyValidPath to FileSystem
