+++ src/libraries/System.Private.CoreLib/src/System/IO/File.cs
@@ -1070,8 +1070,8 @@ namespace System.IO
         /// An I/O error occurred.</exception>
         public static FileSystemInfo CreateSymbolicLink(string path, string pathToTarget)
         {
-            PathInternal.VerifyValidPath(path, nameof(path));
-            PathInternal.VerifyValidPath(pathToTarget, nameof(pathToTarget));
+            FileSystem.VerifyValidPath(path, nameof(path));
+            FileSystem.VerifyValidPath(pathToTarget, nameof(pathToTarget));
             FileSystem.CreateSymbolicLink(path, pathToTarget, isDirectory: false);
             return new FileInfo(path);
         }
@@ -1084,7 +1084,7 @@ namespace System.IO
         /// <returns>A <see cref="FileInfo"/> instance if <paramref name="linkPath"/> exists, independently if the target exists or not. <see langword="null"/> if <paramref name="linkPath"/> does not exist.</returns>
         public static System.IO.FileSystemInfo? ResolveLinkTarget(string linkPath, bool returnFinalTarget = false)
         {
-            PathInternal.VerifyValidPath(linkPath, nameof(linkPath));
+            FileSystem.VerifyValidPath(linkPath, nameof(linkPath));
             return FileSystem.ResolveLinkTarget(linkPath, returnFinalTarget, isDirectory: false);
         }
     }
